ranged enemy

+++ TowerDefence/Assets/Scripts/Tower.cs
@@ -8,7 +8,7 @@ public class Tower : ScriptableObject
     public string towerName;
     public int cost;
     public GameObject tower;
-    public int health = 3;
+    public float health = 3;
 
     
     public int CalculateSellValue(){
